merasip/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pdfplumber==0.11.0
pypdf==4.2.0
reportlab==4.1.0
supabase==2.5.0
python-jose[cryptography]==3.3.0
boto3==1.34.0
httpx==0.27.0
pydantic==2.7.0
python-dotenv==1.0.1
